scraper/requirements.txt
python-dotenv
requests
selenium
bs4
fastapi[standard]
pyotp